+++ scipy/linalg/src/_linalg_lu_det.hh
@@ -131,7 +131,7 @@ inline void ipiv_to_perm(CBLAS_INT *ipiv, CBLAS_INT *perm, CBLAS_INT m, CBLAS_IN
 template<typename T>
 void permute_rows(T *data, const CBLAS_INT *perm, T *tmp, CBLAS_INT m, CBLAS_INT ncols)
 {
-    std::memcpy(tmp, data, m * ncols * sizeof(T));
+    std::memcpy(tmp, data, (size_t)m * ncols * sizeof(T));
     for (CBLAS_INT i = 0; i < m; i++) {
         if (perm[i] != i) {
             std::memcpy(data + i * ncols, tmp + perm[i] * ncols, ncols * sizeof(T));
